put 'core::serde' into its own module file
It started small, but now is large enough to warrant that.
